pyproject.toml
[tool.poetry]
name = "pyladiescon"
version = "0.1.0"
description = ""
authors = ["Ana Dulce Padovan <anadulcepadovan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
django-ninja = "^1.0.1"
django-localflavor = "^4.0"
black = "^23.11.0"
python-slugify = "^8.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
